+++ www/go/core/views/extjs3/usersettings/LookAndFeelPanel.js
@@ -308,68 +308,7 @@ go.usersettings.LookAndFeelPanel = Ext.extend(Ext.Panel, {
 			labelWidth:dp(160),
 			title: t('Notifications','users','core'),
 			items:[
-				this.cbPopupReminders = new Ext.ux.form.XCheckbox({
-					hideLabel: true,
-					boxLabel: t("Show a popup window when a reminder becomes active", "users", "core"),
-					name: 'popup_reminders',
-					listeners: {
-						check: function(cb,checked) {
-							if(checked) {
-								var options = {
-									body: t("Desktop notifications active"),
-									icon: 'views/Extjs3/themes/Group-Office/images/groupoffice.ico'
-								}
-								// Let's check if the browser supports notifications
 
-								if (!("Notification" in window)) {
-									// Browser does not support desktop notification and will show a popup instead
-								} else if (Notification.permission !== 'granted' && (Notification.permission !== 'denied' || Notification.permission === "default")) {
-									Notification.requestPermission(function (permission) {
-									// If the user accepts, let's create a notification
-									try {
-										if (permission === "granted") {
-											var notification = new Notification(t("Desktop notifications active"), options);
-											return true;
-										}
-									} catch(e) {
-										// ignore failure on android
-									}
-									cb.setValue(false);
-
-									});
-								}
-							} 
-						}
-					}
-				}),
-				this.cbPopupEmailNotifications = new Ext.ux.form.XCheckbox({
-					hideLabel: true,
-					boxLabel: t("Show a popup window when an e-mail arrives", "users", "core"),
-					name: 'popup_emails',
-					listeners: {
-						check: function (cb, checked) {
-							if (checked) {
-								var options = {
-									body: t("Desktop notifications active"),
-									icon: 'views/Extjs3/themes/Group-Office/images/groupoffice.ico'
-								}
-								// Let's check if the browser supports notifications
-								if (!("Notification" in window)) {
-									// Browser does not support desktop notification and will show a popup instead
-								} else if (Notification.permission !== 'granted' && (Notification.permission !== 'denied' || Notification.permission === "default")) {
-									Notification.requestPermission(function (permission) {
-										// If the user accepts, let's create a notification
-										if (permission === "granted") {
-											var notification = new Notification(t("Desktop notifications active"), options);
-										} else {
-											cb.setValue(false);
-										}
-									});
-								}
-							}
-						}
-					}
-				}),
 				this.cbEmailReminders = new Ext.ux.form.XCheckbox({
 					hideLabel: true,
 					boxLabel: t("Mail reminders", "users", "core"),